gpu/capture: split the capture file format out of gpu_capture.cpp
Step two, and a PURE MOVE: no line is edited, only relocated. split_file.py proves it --
the two outputs read back from disk rebuild the original 282,267 bytes exactly.

gpu_capture.cpp was 5252 lines doing two jobs. Collecting a frame (intervals, tables,
submit items, the begin/finish/materialize entry points) and writing it to a file (the
byte cursors, the per-field codecs, serialize/deserialize) share a vocabulary but not a
reason to change together. The folder's AGENTS.md described `gpu_capture` as "the capture
itself", which had stopped being one thing.

  gpu_capture.cpp            2409 lines   collection and entry points
  gpu_capture_serialize.cpp  2843 lines   the file format
  gpu_capture_internal.hpp   1399 lines   what both need (previous commit)

The seam came from the reference matrix rather than from reading: the serialization side
is the two entry points, the Writer/Reader byte cursors and the 20 per-field codecs, which
is the group the references pick out. Nothing in it is called from the collection side.

Verified: builds clean; ctest --no-tests=error -j4 reports 457/457 passed, exit 0. Nothing
here changes behaviour, so a green suite is the right check and the byte-identical
reconstruction is the stronger one.

CMakeLists needs no edit -- src/*.cpp is globbed with CONFIGURE_DEPENDS.
